+++ gtaLib/txd.py
@@ -81,6 +81,25 @@ class DeviceType(IntEnum):
     DEVICE_PSP  = 9
 
 #######################################################
+class ImageEncoder:
+    
+    @staticmethod
+    def rgba_to_bgra8888(rgba_data):
+        ret = bytearray()
+        for i in range(0, len(rgba_data), 4):
+            r, g, b, a = rgba_data[i:i+4]
+            ret.extend([b, g, r, a])
+        return bytes(ret)
+    
+    @staticmethod
+    def rgba_to_bgra888(rgba_data):
+        ret = bytearray()
+        for i in range(0, len(rgba_data), 4):
+            r, g, b = rgba_data[i:i+3]
+            ret.extend([b, g, r])
+        return bytes(ret)
+
+#######################################################
 class ImageDecoder:
 
     @staticmethod
